Add support for GHC 9.0.1
The module structure for the GHC API changed in version 9 so this
required using CPP for the imports. Additionally, `RealSrcSpan` was
given an additional parameter so `P.RealSrcSpan _` was changed to
`P.RealSrcSpan {}` to accommodate this.

There was also an issue with some variables that previously had a type
represented by `TyConApp` now have an explicit `ForAllTy`. In
particular `noWay` from `test/Spec.hs` had this issue. This has been
solved with the `getTyCon` helper which looks under `ForAllTy`s. This
should also make it so that polymorphic variables with explicit
`forall`s are not passed over.